fix: navigation within the remote Angular 12 app
First time we try to navigate within the remote app, e.g. from accounts
to account statements, the component doesn't render.

Interestingly, it seems like some kind of race condition because once
loaded, subsequent navigation renders the component without issues as
the lazy loaded module bundle is already downloaded.

A custom web component wrapper with a setTimeout before creating the
element seems to help, but it's unreliable.

So we have gone with a custom event as a way to communicate both apps,
thus leaving remote navigation to be handled entirely by the remote app.