fix: inner router issues
We are fixing three different issues when navigating between the remote
routes and reverting commit 0d64b15b

* In the Angular 12 shell, the previous journey was still there making
another API call when navigating to a new remote route. This was due to
the previous AppComponent being destroyed after the connectRouter of the
new route happened.

* In the Angular 14 shell, when we loaded a new remote route it didn't
show the first time, but it was ok in successive attempts. This was
caused by some kind of race condition with the loadRemoteModule and the
lazy loaded journeys.

* In both apps, we weren't able to use the startWith when defining the
remote routes because the inner router wouldn't recognize the route
change.

Signed-off-by: Pablo Villoslada Puigcerber <pablov@backbase.com>